jcolchon Se controla la carga masiva y la eliminacion de formadores

+++ siga-web/src/main/java/org/itcgae/siga/form/services/impl/FichaCursosServiceImpl.java
@@ -49,6 +49,7 @@ import org.itcgae.siga.db.entities.AdmUsuarios;
 import org.itcgae.siga.db.entities.AdmUsuariosExample;
 import org.itcgae.siga.db.entities.AgeEvento;
 import org.itcgae.siga.db.entities.AgeEventoExample;
+import org.itcgae.siga.db.entities.AgePersonaEvento;
 import org.itcgae.siga.db.entities.CenCliente;
 import org.itcgae.siga.db.entities.CenClienteExample;
 import org.itcgae.siga.db.entities.CenDatoscv;
@@ -308,7 +309,8 @@ public class FichaCursosServiceImpl implements IFichaCursosService {
 				AdmUsuarios usuario = usuarios.get(0);
 				LOGGER.info(
 						"getTrainersCourse() / forPersonacursoExtendsMapper.getTrainers(idInstitucion, idCurso) -> Entrada a forPersonacursoExtendsMapper para obtener los formadores de un curso especifico");
-				formadoresCursoItem = forPersonacursoExtendsMapper.getTrainersCourse(idInstitucion.toString(), idCurso,
+                ForCurso curso = forCursoExtendsMapper.selectByPrimaryKey(new Long(idCurso));
+                formadoresCursoItem = forPersonacursoExtendsMapper.getTrainersCourse(curso.getIdinstitucion().toString(), idCurso,
 						usuario.getIdlenguaje());
 				LOGGER.info(
 						"getTrainersCourse() / forPersonacursoExtendsMapper.getTrainers(idInstitucion, idCurso) -> Salida de forPersonacursoExtendsMapper para obtener los formadores de un curso especifico");
@@ -669,7 +671,6 @@ public class FichaCursosServiceImpl implements IFichaCursosService {
 
 		return record;
 	}
-
 	@Override
 	@Transactional
 	public UpdateResponseDTO deleteTrainersCourse(FormadorCursoDTO formadorCursoDTO, HttpServletRequest request) {
@@ -678,6 +679,7 @@ public class FichaCursosServiceImpl implements IFichaCursosService {
 
 		UpdateResponseDTO updateResponseDTO = new UpdateResponseDTO();
 		Error error = new Error();
+		updateResponseDTO.setStatus(SigaConstants.OK);
 		int response = 1;
 
 		// Conseguimos información del usuario logeado
@@ -698,40 +700,58 @@ public class FichaCursosServiceImpl implements IFichaCursosService {
 				AdmUsuarios usuario = usuarios.get(0);
 
 				try {
-
+					Boolean existePersona = Boolean.FALSE;
 					for (FormadorCursoItem formador : formadorCursoDTO.getFormadoresCursoItem()) {
 
-						ForPersonaCursoExample exampleFormador = new ForPersonaCursoExample();
-						exampleFormador.createCriteria().andIdformadorEqualTo(Long.valueOf(formador.getIdFormador()));
-
-						LOGGER.info(
-								"deleteTrainersCourse() / admUsuariosExtendsMapper.selectByExample() -> Entrada a forPersonacursoExtendsMapper para obtener los formadores de un curso");
-						List<ForPersonaCurso> formadoresList = forPersonacursoExtendsMapper
-								.selectByExample(exampleFormador);
-						LOGGER.info(
-								"deleteTrainersCourse() / admUsuariosExtendsMapper.selectByExample() -> Salida a forPersonacursoExtendsMapper para obtener los formadores de un curso");
-
-						if (null != formadoresList && formadoresList.size() > 0) {
-							ForPersonaCurso formadorDelete = formadoresList.get(0);
-
-							LOGGER.info(
-									"deleteTrainersCourse() / forPersonacursoExtendsMapper.updateByPrimaryKey(notification) -> Entrada a forPersonacursoExtendsMapper para dar de baja a un formador");
-
-							formadorDelete.setFechamodificacion(new Date());
-							formadorDelete.setUsumodificacion(usuario.getIdusuario().longValue());
-							formadorDelete.setFechabaja(new Date());
-							forPersonacursoExtendsMapper.updateByPrimaryKey(formadorDelete);
-
-							LOGGER.info(
-									"deleteTrainersCourse() / forPersonacursoExtendsMapper.updateByPrimaryKey(notification) -> Salida a forPersonacursoExtendsMapper para dar de baja a un formador");
-
+						//Comprobamos que el formador no esté asignado a alguna sesión
+						
+						List<AgePersonaEvento> personaSesion= forPersonacursoExtendsMapper.selectSesionesFormador(formador);
+						if (null != personaSesion && personaSesion.size()>0) {
+							response = 0;
+							error.setCode(400);
+							error.setDescription("Los formadores a eliminar no pueden estar asignados a ninguna sesión");
+							existePersona = Boolean.TRUE;
+							updateResponseDTO.setStatus(SigaConstants.KO);
+							break;
+							
 						}
-					}
-
+					}	
+						if (!existePersona) {
+						
+							for (FormadorCursoItem formador : formadorCursoDTO.getFormadoresCursoItem()) {
+								ForPersonaCursoExample exampleFormador = new ForPersonaCursoExample();
+								exampleFormador.createCriteria().andIdformadorEqualTo(Long.valueOf(formador.getIdFormador()));
+		
+								
+								LOGGER.info(
+										"deleteTrainersCourse() / admUsuariosExtendsMapper.selectByExample() -> Entrada a forPersonacursoExtendsMapper para obtener los formadores de un curso");
+								List<ForPersonaCurso> formadoresList = forPersonacursoExtendsMapper
+										.selectByExample(exampleFormador);
+								LOGGER.info(
+										"deleteTrainersCourse() / admUsuariosExtendsMapper.selectByExample() -> Salida a forPersonacursoExtendsMapper para obtener los formadores de un curso");
+		
+								if (null != formadoresList && formadoresList.size() > 0) {
+									ForPersonaCurso formadorDelete = formadoresList.get(0);
+		
+									LOGGER.info(
+											"deleteTrainersCourse() / forPersonacursoExtendsMapper.updateByPrimaryKey(notification) -> Entrada a forPersonacursoExtendsMapper para dar de baja a un formador");
+		
+									formadorDelete.setFechamodificacion(new Date());
+									formadorDelete.setUsumodificacion(usuario.getIdusuario().longValue());
+									formadorDelete.setFechabaja(new Date());
+									forPersonacursoExtendsMapper.updateByPrimaryKey(formadorDelete);
+		
+									LOGGER.info(
+											"deleteTrainersCourse() / forPersonacursoExtendsMapper.updateByPrimaryKey(notification) -> Salida a forPersonacursoExtendsMapper para dar de baja a un formador");
+								}
+							
+							}
+						}	
 				} catch (Exception e) {
 					response = 0;
 					error.setCode(400);
 					error.setDescription("Se ha producido un error en BBDD contacte con su administrador");
+					updateResponseDTO.setStatus(SigaConstants.KO);
 				}
 			}
 		}
@@ -745,7 +765,6 @@ public class FichaCursosServiceImpl implements IFichaCursosService {
 		updateResponseDTO.setError(error);
 		return updateResponseDTO;
 	}
-
 	@Override
 	@Transactional
 	public InsertResponseDTO saveCourse(CursoItem cursoItem, HttpServletRequest request) {
@@ -1686,7 +1705,7 @@ public class FichaCursosServiceImpl implements IFichaCursosService {
 			List<AdmUsuarios> usuarios = admUsuariosExtendsMapper.selectByExample(exampleUsuarios);
 			LOGGER.info(
 					"uploadFileExcel() / admUsuariosExtendsMapper.selectByExample() -> Salida de admUsuariosExtendsMapper para obtener información del usuario logeado");
-
+			ForCurso curso = forCursoExtendsMapper.selectByPrimaryKey(Long.parseLong(String.valueOf(idCurso)));
 			if (null != usuarios && usuarios.size() > 0) {
 				AdmUsuarios usuario = usuarios.get(0);
 
@@ -1700,6 +1719,35 @@ public class FichaCursosServiceImpl implements IFichaCursosService {
 					}
 				}
 
+				// Generamos el fichero de errores
+				byte[] bytesLog = ExcelHelper.createExcelBytes(SigaConstants.CAMPOSPLOGCURSO, datosLog);
+
+				// Insertamos en tabla For_inscripcion_Masiva el fichero que se ha parseado
+				forInscripcionesmasivas.setIdinstitucion(curso.getIdinstitucion());
+				forInscripcionesmasivas.setNombrefichero(nombreFichero);
+				forInscripcionesmasivas.setFechamodificacion(new Date());
+				forInscripcionesmasivas.setUsumodificacion(Long.valueOf(usuario.getIdusuario()));
+				forInscripcionesmasivas.setIdcurso(Long.valueOf(idCurso));
+				String numLineas = String.valueOf(inscripcionList.size());
+				forInscripcionesmasivas.setNumerolineastotales(Long.valueOf(numLineas));
+				String lineasCorrectas = String.valueOf(inscripcionList.size() - registrosErroneos);
+				forInscripcionesmasivas.setInscripcionescorrectas(Long.valueOf(lineasCorrectas));
+
+	
+				
+				Long idFile = uploadFile(file.getBytes(), forInscripcionesmasivas, false, usuario);
+				Long idLogFile = uploadFile(bytesLog, forInscripcionesmasivas, true, usuario);
+				
+				forInscripcionesmasivas.setIdfichero(idFile);
+				forInscripcionesmasivas.setIdficherolog(idLogFile);
+				LOGGER.info(
+						"uploadFileExcel() / forInscripcionesmasivasMapper.insert(forInscripcionesmasivas) -> Entrada a forInscripcionesmasivasMapper para insertar el fichero parseado carga masiva de inscripciones");
+
+				int result = forInscripcionesmasivasMapper.insert(forInscripcionesmasivas);
+
+				LOGGER.info(
+						"uploadFileExcel() / forInscripcionesmasivasMapper.insert(forInscripcionesmasivas) -> Salida a forInscripcionesmasivasMapper para insertar el fichero parseado carga masiva de inscripciones");
+
 
 				// Si no hay errores se insertan las inscripciones nuevas
 				if (registrosErroneos == 0) {
@@ -1712,7 +1760,7 @@ public class FichaCursosServiceImpl implements IFichaCursosService {
 							inscripcionInsert.setFechamodificacion(new Date());
 							inscripcionInsert.setIdcurso(Long.valueOf(inscripcion.getIdCurso()));
 							inscripcionInsert.setIdestadoinscripcion(SigaConstants.INSCRIPCION_PENDIENTE);
-							inscripcionInsert.setIdinstitucion(idInstitucion);
+							inscripcionInsert.setIdinstitucion(curso.getIdinstitucion());
 							inscripcionInsert.setIdpersona(inscripcion.getIdPersona());
 							inscripcionInsert.setUsumodificacion(usuario.getIdusuario().longValue());
 							inscripcionInsert.setIdcargainscripcion(forInscripcionesmasivas.getIdcargainscripcion());
@@ -1720,7 +1768,7 @@ public class FichaCursosServiceImpl implements IFichaCursosService {
 							LOGGER.info(
 									"uploadFileExcel() / forInscripcionExtendsMapper.insert(inscripcionInsert) -> Entrada a forInscripcionExtendsMapper para insertar una inscripcion");
 
-							int result = forInscripcionExtendsMapper.insert(inscripcionInsert);
+							result = forInscripcionExtendsMapper.insert(inscripcionInsert);
 
 							LOGGER.info(
 									"uploadFileExcel() / forInscripcionExtendsMapper.insert(inscripcionInsert) -> Salida a forInscripcionExtendsMapper para insertar una inscripcion");
@@ -1754,35 +1802,6 @@ public class FichaCursosServiceImpl implements IFichaCursosService {
 					error.setCode(400);
 					error.setDescription("Hay errores en las inscripciones subidas");
 				}
-
-				// Generamos el fichero de errores
-				byte[] bytesLog = ExcelHelper.createExcelBytes(SigaConstants.CAMPOSPLOGCURSO, datosLog);
-
-				// Insertamos en tabla For_inscripcion_Masiva el fichero que se ha parseado
-				forInscripcionesmasivas.setIdinstitucion(idInstitucion);
-				forInscripcionesmasivas.setNombrefichero(nombreFichero);
-				forInscripcionesmasivas.setFechamodificacion(new Date());
-				forInscripcionesmasivas.setUsumodificacion(Long.valueOf(usuario.getIdusuario()));
-				forInscripcionesmasivas.setIdcurso(Long.valueOf(idCurso));
-				String numLineas = String.valueOf(inscripcionList.size());
-				forInscripcionesmasivas.setNumerolineastotales(Long.valueOf(numLineas));
-				String lineasCorrectas = String.valueOf(inscripcionList.size() - registrosErroneos);
-				forInscripcionesmasivas.setInscripcionescorrectas(Long.valueOf(lineasCorrectas));
-
-	
-				
-				Long idFile = uploadFile(file.getBytes(), forInscripcionesmasivas, false, usuario);
-				Long idLogFile = uploadFile(bytesLog, forInscripcionesmasivas, true, usuario);
-				
-				forInscripcionesmasivas.setIdfichero(idFile);
-				forInscripcionesmasivas.setIdficherolog(idLogFile);
-				LOGGER.info(
-						"uploadFileExcel() / forInscripcionesmasivasMapper.insert(forInscripcionesmasivas) -> Entrada a forInscripcionesmasivasMapper para insertar el fichero parseado carga masiva de inscripciones");
-
-				int result = forInscripcionesmasivasMapper.insert(forInscripcionesmasivas);
-
-				LOGGER.info(
-						"uploadFileExcel() / forInscripcionesmasivasMapper.insert(forInscripcionesmasivas) -> Salida a forInscripcionesmasivasMapper para insertar el fichero parseado carga masiva de inscripciones");
 
 
 			}
@@ -2564,7 +2583,7 @@ public class FichaCursosServiceImpl implements IFichaCursosService {
 	
 									} else {
 										error.setDescription(
-												"Las inscripciones no se pueden aprobar porque no estÁn en estado pendiente.");
+												"Las inscripciones no se pueden aprobar porque no están en estado pendiente.");
 	
 									}
 								}
